Moved all python codes into UTIL. Collected tests into run_test.sh. Updated readme
