+++ README.md
@@ -6,5 +6,7 @@ This repository documents the activity of 2017 summer students in Orengo's group
 See **ama_entries.log** for ambiguous/amazing entries.
 [Acornyms (CATH_trac, internal)](http://trac.cathdb.info/projects/cath/wiki/DomChop) 
 
+See **util/** for ongoing python codes related to domchop.
+
 #### CATH workflow
 ![CATH pipeline flowchart](http://update.cathdb.info/images/flow_stage_diagram.jpg)
