+++ run_tests.sh
@@ -1,2 +1,10 @@
-#nosetests tests; echo n $?
-python -m tests.toy_test
+cd "$(dirname "$0")" ### set directory to the same as this script
+cd tests
+for f in *.py
+do
+echo running $f
+# python $f > $f.log
+python $f
+
+done
+echo 'All tests done'
