+++ util/README.md
@@ -1,0 +1,14 @@
+
+
+To configure the modeller temporarily, run this script to set the environment variables temporarily. You need to **edit** the script according to your installation!
+```
+source config_modeller.sh
+```
+
+After configuration, please run tests with 
+```
+chmod +x run_tests.sh # To make the script executable
+./run_tests.sh ## To run the tests
+```
+
+
