+++ util/pdbs/2odlA_chop_cath
