+++ util/tests/ref_DOPEs.csv
@@ -1,3 +1,5 @@
+4ufqA,2.0148097850445352
 2odlA_chop_cath,-2.019600620618391
 2odlA_sup,6.0432001461323965
+4ufq.pdb1,1.994125356634698
 4xz8A_chop,-1.6751284312435732
